services/progress_service/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "progress-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "httpx>=0.27",
  "alembic>=1.13",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "passlib[argon2]>=1.7",
  "argon2-cffi>=23.1.0",
  "pyjwt[crypto]>=2.8",
  "psycopg[binary]",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "ruff",
  "black",
  "mypy",
  "httpx",
  "alembic",
  "pre-commit",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = ["app", "pyproject.toml"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["app"]
indent-width = 2
extend-exclude = [
  "alembic",
  "migrations",
  ".venv",
  "venv",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = [
  "E501",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = true

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_unused_ignores = true
ignore_missing_imports = true
disallow_incomplete_defs = true
disallow_untyped_defs = false
disallow_untyped_calls = false
check_untyped_defs = true
warn_redundant_casts = true
warn_no_return = true
strict_optional = true
pretty = true
show_error_context = true
show_column_numbers = true
exclude = [
    "alembic",
    "migrations",
    ".venv",
    "venv"
]
explicit_package_bases = true
mypy_path = "."
files = ["app"]
